feat: add simple __version__ test
